Move classes to a new namespace
