+++ src/Borm/Data/InternalTransaction.cs
@@ -1,4 +1,5 @@
 ﻿using System.Diagnostics;
+using Borm.Data.Storage;
 using Borm.Properties;
 
 namespace Borm.Data;
