+++ src/Borm/Data/Storage/EntityMaterializer.cs
@@ -2,7 +2,7 @@
 using Borm.Model.Metadata;
 using Borm.Properties;
 
-namespace Borm.Data;
+namespace Borm.Data.Storage;
 
 internal sealed class EntityMaterializer
 {
