+++ src/Borm/Data/Storage/Table.cs
@@ -7,7 +7,7 @@ using Borm.Model;
 using Borm.Model.Metadata;
 using Borm.Properties;
 
-namespace Borm.Data;
+namespace Borm.Data.Storage;
 
 [DebuggerTypeProxy(typeof(TableDebugView))]
 internal sealed class Table
