+++ src/Borm/Data/Storage/TableGraph.cs
@@ -1,4 +1,4 @@
-﻿namespace Borm.Data;
+﻿namespace Borm.Data.Storage;
 
 internal sealed class TableGraph
 {
