+++ src/Borm/DataContext.cs
@@ -1,5 +1,6 @@
 ﻿using System.Diagnostics;
 using Borm.Data;
+using Borm.Data.Storage;
 using Borm.Model;
 using Borm.Model.Metadata;
 using Borm.Properties;
