+++ src/Borm/Model/Metadata/EntityMaterializationBinding.cs
@@ -2,7 +2,7 @@
 using System.Diagnostics.CodeAnalysis;
 using System.Linq.Expressions;
 using System.Reflection;
-using Borm.Data;
+using Borm.Data.Storage;
 using Borm.Extensions;
 
 namespace Borm.Model.Metadata;
