+++ tests/Borm.Tests/Data/Storage/ChangeTest.cs
@@ -1,8 +1,8 @@
-﻿using Borm.Data;
+﻿using Borm.Data.Storage;
 using static Borm.Tests.Mocks.TableMocks;
 using static Borm.Tests.Mocks.ValueBufferMockHelper;
 
-namespace Borm.Tests.Data;
+namespace Borm.Tests.Data.Storage;
 
 public sealed class ChangeTest
 {
