+++ tests/Borm.Tests/Data/Storage/EntityMaterializerTest.cs
@@ -1,9 +1,9 @@
-﻿using Borm.Data;
+﻿using Borm.Data.Storage;
 using static Borm.Tests.Mocks.EntityMetadataMocks;
 using static Borm.Tests.Mocks.TableMocks;
 using static Borm.Tests.Mocks.ValueBufferMockHelper;
 
-namespace Borm.Tests.Data;
+namespace Borm.Tests.Data.Storage;
 
 public sealed class EntityMaterializerTest
 {
