+++ tests/Borm.Tests/Mocks/EntityMetadataMocks.cs
@@ -1,4 +1,4 @@
-﻿using Borm.Data;
+﻿using Borm.Data.Storage;
 using Borm.Model;
 using Borm.Model.Metadata;
